requirements.txt
asgiref==3.2.3
backcall==0.1.0
colorama==0.4.3
decorator==4.4.1
Django==3.0.3
-e git+https://github.com/jazzband/django-debug-toolbar.git@bad48b28e74618c2d21453202b0ed59cdb9fd0b3#egg=django_debug_toolbar
ipdb==0.12.3
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
parso==0.6.1
pickleshare==0.7.5
prompt-toolkit==3.0.3
Pygments==2.5.2
python-slugify==4.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.3
wcwidth==0.1.8
